requirements.txt
async-generator==1.8
attrs==17.2.0
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
protobuf==3.4.0
requests==2.18.4
six==1.10.0
sortedcontainers==1.5.7
git+https://github.com/python-trio/trio
urllib3==1.22
